remove obsolete files and move schema to separate folder
